# apps/api-client-go/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@boxlite.com
    name: BoxLite Platforms Inc.
    url: https://www.boxlite.io
  description: BoxLite AI platform API Docs
  title: BoxLite
  version: "1.0"
servers:
- url: http://localhost:3000
paths:
...
  /organization/{organizationId}/billing/payment:
    get:
      operationId: BillingPaymentController_getPaymentState
      parameters:
      - explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      tags:
      - billing
  /organization/{organizationId}/billing/payment/setup:
    post:
      operationId: BillingPaymentController_setupPaymentMethod
      parameters:
      - explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "201":
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      tags:
      - billing
  /organization/{organizationId}/billing/auto-reload:
    put:
      operationId: BillingPaymentController_setAutoReload
      parameters:
      - explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      tags:
      - billing
  /organization/{organizationId}/billing/top-ups:
    post:
      operationId: BillingPaymentController_createTopUp
      parameters:
      - explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      - description: Stable client-generated key reused when retrying the same top-up
          request
        explode: false
        in: header
        name: idempotency-key
        required: true
        schema:
          type: string
        style: simple
      responses:
        "201":
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      tags:
      - billing
  /organization/{organizationId}/billing/receipts:
    get:
      operationId: BillingPaymentController_listReceipts
      parameters:
      - explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      - explode: true
        in: query
        name: page
        required: true
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: pageSize
        required: true
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: query
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      tags:
      - billing
  /billing/webhooks/payment:
    post:
      operationId: PaymentWebhookController_handle
      parameters:
      - explode: false
        in: header
        name: stripe-signature
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: ""
      tags:
      - billing
...
components:
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      description: API Key access
      scheme: bearer
      type: http
    oauth2:
      openIdConnectUrl: http://localhost:3000/.well-known/openid-configuration
      type: openIdConnect